Extract XPath and CSS classes from selector builders
