+++ lib/watir-webdriver/locators.rb
@@ -4,6 +4,7 @@ require 'watir-webdriver/locators/element/validator'
 
 require 'watir-webdriver/locators/button/locator'
 require 'watir-webdriver/locators/button/selector_builder'
+require 'watir-webdriver/locators/button/selector_builder/xpath'
 require 'watir-webdriver/locators/button/validator'
 
 require 'watir-webdriver/locators/cell/locator'
@@ -17,3 +18,4 @@ require 'watir-webdriver/locators/text_area/selector_builder'
 
 require 'watir-webdriver/locators/text_field/locator'
 require 'watir-webdriver/locators/text_field/selector_builder'
+require 'watir-webdriver/locators/text_field/selector_builder/xpath'
