+++ lib/watir-webdriver/locators/text_field/selector_builder/xpath.rb
@@ -1,0 +1,15 @@
+module Watir
+  class TextField
+    class SelectorBuilder
+      class XPath < Element::SelectorBuilder::XPath
+        def lhs_for(building, key)
+          if building == :input && key == :text
+            "@value"
+          else
+            super
+          end
+        end
+      end
+    end
+  end
+end
